requirements.txt
tensorflow>=2.15.0
torch>=2.1.0
torchvision>=0.16.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
streamlit>=1.28.0
sqlalchemy>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
gradio>=4.0.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
xformers>=0.0.22
